Allow FlexDLL objects outside of flexdir
It is somewhat strange having to install .obj and .o files in a BINDIR
when a LIBDIR is available.

Old behaviour is preserved - so flexlink will always prefer flexdll_*.$(O)
and flexdll_initer_*.$(O) from the same directory as flexlink, but if they
are missing then the search path is used, as for any other object file.

The default manifests must still be in the same directory as flexlink.

(flexdll-commit: ec8e5fc2bef497c80e9c1f845684c5ef291cb59b)