Remove dash from mingw prefixes.
Patch by whitequark, who wrote:

"there are two generally
accepted ways to locate compilers. One is to specify
the triple, in which case the dash between triple and tool
name is implicit and the tool is searched in PATH.
Another is to specify the complete prefix to the compiler,
in which case the prefix is prepended to the tool name
without anything before it. This second way allows the most
flexibility: you can have the tool be invoked as any of:
      - ld
      - x86_64-w64-mingw32-ld
      - /usr/local/bin/x86_64-w64-mingw32-ld
    etc.
Implicitly adding a dash goes against the established convention
and makes the mechanism less general."

(flexdll-commit: 6af2d26acab60f61dbe5716f01e24d42b42fd370)