Ensure commands always silent for MSVC
If a command is routed through bash because it's too long, the fact it's
meant to be piped to NUL is forgotten. Ensure the bash command is piped
to >/dev/null (with the slightly tedious recovery of the script needed
if the command fails).

(flexdll-commit: 09a34ad3f885fca5b8b1d6ce18f6062a24a2581c)