Correct the parsing of /n COFF lib headers
COFF defines any archive header beginning with / as special. Before, we
assumed that if the header was neither / nor // then it must be /n and
parsed accordingly. That breaks with the new /<XFGHASH>/ header. Harden
the parsing code to allow these new header to be correctly skipped.

(flexdll-commit: ab569cf3c171af2f29029687209ac9000de8f257)