Store the version number in one place
The VERSION macro in Makefile is now the single source of the version of
FlexDLL. This is used when compiling version.rc.

version.rc is corrected to adopt Windows version format, so the current
release is rendered as 0.43.0.0 instead of 0.0.0.43.

flexdll.opam is now the primary source for the version number for the
package, and both version.ml and version.rc are compiled using the
number taken from there.

(flexdll-commit: c5d59b5fc0812da927b8c12b2211c1acd3026309)