Fix stack reserve to read from the COFF file header
The COFF File Header (Object and Image) [1] starts just after the
signature "PE\0\0". The `read ic 0 20` seeks the offset 0 (the start)
in the PE file, instead of the offset of the header. This leads to
incorrectly reading and writing values in the PE headers.

Fix reading the file address of new exe header, a LONG instead of a
WORD. Thanks to Nicolás Ojeda Bär.

Remove outdated debugging comment: it's printing a int32, but the
field is actually an uint32 on PE32 and an uint64 on PE32+.

[1]: https://learn.microsoft.com/en-us/windows/win32/debug/pe-format#coff-file-header-object-and-image

(flexdll-commit: 98e6591ba20225a70441b54f1d5350d225eef8c7)